Optimize update import and add tags column in node_changset
During update do not issue a separate SELECT to find out if the
data exists but simply try an UPDATE. If it fails, do an INSERT.
Also clean up the code slightly on the way.

The new tags column in node_changeset is a hstore with the new
tags of the node. That should allow for a more efficient search
for updated nodes, while avoiding the need for an expensive
index on the tags column in the nodes table.